Show debug SQL statements ready for copy-pasting
With the first line of the SQL on the first line, the file name (e.g. `base.py:123`) is copy-pasted into SQL and requirees manual editing. With this fix, the whole statement will be ready for copy-pasting without any fixes and cleanups. The statements are often multi-line anyway, so it does not damage the logging.