Remove unused code: schema fetching via dialects
The schemas are currently fetching via `Database.select_table_schema()` & `Database.query_table_schema()`, so this dialectic code is unused.

It is also questionable that the schema fetching must be done via dialects at all: some databases can provide an API (e.g. HTTP API) to fetch the database metadata, such as Databricks. So it belongs to the database, not to dialects.